code splitted into different variants instead using an if-else
